refactor(libwiki): extract shared budget counters into budget.js
Move countLines/countWords out of audit/scopes.js into a new budget.js so
the audit and the rotation primitive's bisecting seal (spec 1450) import one
canonical pair and cannot drift.
