+++ libraries/libwiki/src/index.js
@@ -35,5 +35,6 @@ export {
   appendEntry,
 } from "./weekly-log.js";
 export { buildDigest } from "./boot.js";
+export { countLines, countWords } from "./budget.js";
 export { RULES } from "./audit/rules.js";
 export { resolveScope as resolveAuditScope } from "./audit/scopes.js";
